+++ src/test/kotlin/com/contentful/java/cma/SpaceTests.kt
@@ -27,25 +27,25 @@ import kotlin.test.assertTrue
 import org.junit.Test as test
 
 class SpaceTests : BaseTest() {
-    test fun testCreate() {
+    @test fun testCreate() {
         val requestBody = TestUtils.fileToString("space_create_request.json")
         val responseBody = TestUtils.fileToString("space_create_response.json")
         server!!.enqueue(MockResponse().setResponseCode(200).setBody(responseBody))
 
         val result = assertTestCallback(client!!.spaces().async().create(
-                "xxx", TestCallback()) as TestCallback)
+                "xxx", TestCallback()) as TestCallback)!!
 
-        assertEquals("spaceid", result.getSys()["id"])
-        assertEquals("xxx", result.getName())
+        assertEquals("spaceid", result.sys["id"])
+        assertEquals("xxx", result.name)
 
         // Request
         val recordedRequest = server!!.takeRequest()
-        assertEquals("POST", recordedRequest.getMethod())
-        assertEquals("/spaces", recordedRequest.getPath())
-        assertEquals(requestBody, recordedRequest.getUtf8Body())
+        assertEquals("POST", recordedRequest.method)
+        assertEquals("/spaces", recordedRequest.path)
+        assertEquals(requestBody, recordedRequest.utf8Body)
     }
 
-    test fun testCreateInOrg() {
+    @test fun testCreateInOrg() {
         server!!.enqueue(MockResponse().setResponseCode(200).setBody("{}"))
 
         assertTestCallback(client!!.spaces().async().create(
@@ -53,12 +53,12 @@ class SpaceTests : BaseTest() {
 
         // Request
         val recordedRequest = server!!.takeRequest()
-        assertEquals("POST", recordedRequest.getMethod())
-        assertEquals("/spaces", recordedRequest.getPath())
+        assertEquals("POST", recordedRequest.method)
+        assertEquals("/spaces", recordedRequest.path)
         assertEquals("org", recordedRequest.getHeader("X-Contentful-Organization"))
     }
 
-    test fun testDelete() {
+    @test fun testDelete() {
         server!!.enqueue(MockResponse().setResponseCode(200))
 
         assertTestCallback(client!!.spaces().async().delete(
@@ -66,53 +66,53 @@ class SpaceTests : BaseTest() {
 
         // Request
         val recordedRequest = server!!.takeRequest()
-        assertEquals("DELETE", recordedRequest.getMethod())
-        assertEquals("/spaces/spaceid", recordedRequest.getPath())
+        assertEquals("DELETE", recordedRequest.method)
+        assertEquals("/spaces/spaceid", recordedRequest.path)
     }
 
-    test fun testFetchAll() {
+    @test fun testFetchAll() {
         val responseBody = TestUtils.fileToString("space_fetch_all_response.json")
         server!!.enqueue(MockResponse().setResponseCode(200).setBody(responseBody))
 
         val result = assertTestCallback(client!!.spaces().async()
-                .fetchAll(TestCallback()) as TestCallback)
+                .fetchAll(TestCallback()) as TestCallback)!!
 
-        val items = result.getItems()
-        assertEquals("Array", result.getSys()["type"])
-        assertEquals(2, items.size())
-        assertEquals(2, result.getTotal())
-        assertEquals(0, result.getSkip())
-        assertEquals(25, result.getLimit())
+        val items = result.items
+        assertEquals("Array", result.sys["type"])
+        assertEquals(2, items.size)
+        assertEquals(2, result.total)
+        assertEquals(0, result.skip)
+        assertEquals(25, result.limit)
 
         // Space #1
-        var sys = items[0].getSys()
+        var sys = items[0].sys
         assertEquals("Space", sys["type"])
         assertEquals("id1", sys["id"])
         assertEquals(1.toDouble(), sys["version"])
         assertEquals("2014-03-21T08:43:52Z", sys["createdAt"])
         assertEquals("2014-04-27T18:16:10Z", sys["updatedAt"])
-        assertEquals("space1", items[0].getName())
+        assertEquals("space1", items[0].name)
 
         // Created By
-        var map = (items[0].getSys()["createdBy"] as Map<*, *>)["sys"] as Map<*, *>
+        var map = (items[0].sys["createdBy"] as Map<*, *>)["sys"] as Map<*, *>
         assertEquals("Link", map["type"])
         assertEquals("User", map["linkType"])
         assertEquals("user1", map["id"])
 
         // Updated By
-        map = (items[0].getSys()["updatedBy"] as Map<*, *>)["sys"] as Map<*, *>
+        map = (items[0].sys["updatedBy"] as Map<*, *>)["sys"] as Map<*, *>
         assertEquals("Link", map["type"])
         assertEquals("User", map["linkType"])
         assertEquals("user1", map["id"])
 
         // Space #2
-        sys = items[1].getSys()
+        sys = items[1].sys
         assertEquals("Space", sys["type"])
         assertEquals("id2", sys["id"])
         assertEquals(2.toDouble(), sys["version"])
         assertEquals("2014-05-19T09:00:27Z", sys["createdAt"])
         assertEquals("2014-07-09T07:48:24Z", sys["updatedAt"])
-        assertEquals("space2", items[1].getName())
+        assertEquals("space2", items[1].name)
 
         // Created By
         map = (sys["createdBy"] as Map<*, *>)["sys"] as Map<*, *>
@@ -128,24 +128,24 @@ class SpaceTests : BaseTest() {
 
         // Request
         val request = server!!.takeRequest()
-        assertEquals("GET", request.getMethod())
-        assertEquals("/spaces", request.getPath())
+        assertEquals("GET", request.method)
+        assertEquals("/spaces", request.path)
     }
 
-    test fun testFetchWithId() {
+    @test fun testFetchWithId() {
         val responseBody = TestUtils.fileToString("space_fetch_one_response.json")
         server!!.enqueue(MockResponse().setResponseCode(200).setBody(responseBody))
 
         val result = assertTestCallback(client!!.spaces().async().fetchOne(
-                "spaceid", TestCallback()) as TestCallback)
+                "spaceid", TestCallback()) as TestCallback)!!
 
-        val sys = result.getSys()
+        val sys = result.sys
         assertEquals("Space", sys["type"])
         assertEquals("id1", sys["id"])
         assertEquals(1.toDouble(), sys["version"])
         assertEquals("2014-03-21T08:43:52Z", sys["createdAt"])
         assertEquals("2014-04-27T18:16:10Z", sys["updatedAt"])
-        assertEquals("space1", result.getName())
+        assertEquals("space1", result.name)
 
         // Created By
         var map = (sys["createdBy"] as Map<*, *>)["sys"] as Map<*, *>
@@ -161,55 +161,55 @@ class SpaceTests : BaseTest() {
 
         // Request
         val request = server!!.takeRequest()
-        assertEquals("GET", request.getMethod())
-        assertEquals("/spaces/spaceid", request.getPath())
+        assertEquals("GET", request.method)
+        assertEquals("/spaces/spaceid", request.path)
     }
 
-    test fun testFetchLocales() {
+    @test fun testFetchLocales() {
         val responseBody = TestUtils.fileToString("space_fetch_locales_response.json")
         server!!.enqueue(MockResponse().setResponseCode(200).setBody(responseBody))
 
         val result = assertTestCallback(client!!.spaces().async().fetchLocales(
-                "spaceid", TestCallback()) as TestCallback)
+                "spaceid", TestCallback()) as TestCallback)!!
 
-        val item = result.getItems()[0]
-        assertEquals("U.S. English", item.getName())
-        assertEquals("en-US", item.getCode())
-        assertTrue(item.isDefault())
-        assertTrue(item.isPublished())
+        val item = result.items[0]
+        assertEquals("U.S. English", item.name)
+        assertEquals("en-US", item.code)
+        assertTrue(item.isDefault)
+        assertTrue(item.isPublished)
 
         // Request
         val recordedRequest = server!!.takeRequest()
-        assertEquals("GET", recordedRequest.getMethod())
-        assertEquals("/spaces/spaceid/locales", recordedRequest.getPath())
+        assertEquals("GET", recordedRequest.method)
+        assertEquals("/spaces/spaceid/locales", recordedRequest.path)
     }
 
-    test fun testUpdate() {
+    @test fun testUpdate() {
         val requestBody = TestUtils.fileToString("space_update_request.json")
         val responseBody = TestUtils.fileToString("space_update_response.json")
         server!!.enqueue(MockResponse().setResponseCode(200).setBody(responseBody))
 
         var space = gson!!.fromJson(
                 TestUtils.fileToString("space_update_object.json"),
-                javaClass<CMASpace>())
+                CMASpace::class.java)
 
         space.setName("newname")
 
         space = assertTestCallback(client!!.spaces().async().update(
                 space, TestCallback()) as TestCallback)
 
-        assertEquals(2.toDouble(), space.getSys()["version"])
-        assertEquals("newname", space.getName())
+        assertEquals(2.toDouble(), space.sys["version"])
+        assertEquals("newname", space.name)
 
         // Request
         val recordedRequest = server!!.takeRequest()
-        assertEquals("PUT", recordedRequest.getMethod())
-        assertEquals("/spaces/spaceid", recordedRequest.getPath())
+        assertEquals("PUT", recordedRequest.method)
+        assertEquals("/spaces/spaceid", recordedRequest.path)
         assertNotNull(recordedRequest.getHeader("X-Contentful-Version"))
-        assertEquals(requestBody, recordedRequest.getUtf8Body())
+        assertEquals(requestBody, recordedRequest.utf8Body)
     }
 
-    test(expected = Exception::class)
+    @test(expected = Exception::class)
     fun testUpdateFailsWithoutVersion() {
         ModuleTestUtils.assertUpdateWithoutVersion {
             val space: CMASpace = CMASpace().setName("name").setId("id")